third_party/blink/web_tests/platform/generic/fast/forms/formmethod-attribute-test-expected.txt

Test formmethod attribute

    
Missing value
PASS ip1.formMethod is ""
PASS ip1.getAttribute('formmethod') is null


Empty value
PASS ip2.formMethod is "get"
PASS ip2.getAttribute('formmethod') is ""


Post value
PASS ip3.formMethod is "post"
PASS ip3.getAttribute('formmethod') is "post"


Get value
PASS ip4.formMethod is "get"
PASS ip4.getAttribute('formmethod') is "get"


Invalid value
PASS ip5.formMethod is "get"
PASS ip5.getAttribute('formmethod') is "foo"


PASS successfullyParsed is true

TEST COMPLETE
